requirements.txt
amqp==2.5.2
app-version==1.0.1
appnope==0.1.0
asgiref==3.2.3
backcall==0.2.0
beautifulsoup4==4.8.2
billiard==3.6.3.0
celery==4.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
DAWG-Python==0.7.2
decorator==4.4.2
Django==3.0.1
django-bootstrap4==1.1.1
django-braces==1.14.0
django-ckeditor==6.0.0
django-crispy-forms==1.8.1
django-js-asset==1.2.2
django-scheduler==0.9.3
django-tinymce==2.8.0
docopt==0.6.2
docxtpl==0.10.0
icalendar==4.0.7
idna==2.10
importlib-metadata==1.7.0
install==1.3.3
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
kombu==4.6.8
lxml==4.5.2
MarkupSafe==1.1.1
notify==0.3.1
numpy==1.19.4
pandas==1.1.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pika==1.1.0
Pillow==6.2.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.1
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.404381.4453942
python-dateutil==2.8.1
python-docx==0.8.10
python-git==2018.2.1
pytz==2019.3
requests==2.24.0
Send2Trash==1.5.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
traitlets==5.0.5
urllib3==1.25.10
validate-email==1.3
vine==1.3.0
wcwidth==0.2.5
zipp==3.1.0
